external_tools/pet4dattes/requirements.txt
odfpy
openpyxl
pandas
xlrd
